wip: refactor directory structure
