+++ packages/@n8n/ai-workflow-builder.ee/evaluations/programmatic/evaluators/agent-prompt.ts
@@ -1,1 +1,1 @@
-export * from '../../../src/programmatic/evaluators/agent-prompt';
+export * from '@/validation/evaluators/agent-prompt';
