+++ packages/@n8n/ai-workflow-builder.ee/evaluations/programmatic/evaluators/connections.ts
@@ -1,1 +1,1 @@
-export * from '../../../src/programmatic/evaluators/connections';
+export * from '@/validation/evaluators/connections';
