+++ packages/@n8n/ai-workflow-builder.ee/evaluations/programmatic/evaluators/from-ai.ts
@@ -1,1 +1,1 @@
-export * from '../../../src/programmatic/evaluators/from-ai';
+export * from '@/validation/evaluators/from-ai';
