+++ packages/@n8n/ai-workflow-builder.ee/evaluations/programmatic/evaluators/tools.ts
@@ -1,1 +1,1 @@
-export * from '../../../src/programmatic/evaluators/tools';
+export * from '@/validation/evaluators/tools';
