+++ packages/@n8n/ai-workflow-builder.ee/evaluations/programmatic/evaluators/trigger.ts
@@ -1,1 +1,1 @@
-export * from '../../../src/programmatic/evaluators/trigger';
+export * from '@/validation/evaluators/trigger';
