+++ packages/@n8n/ai-workflow-builder.ee/evaluations/utils/evaluation-reporter.ts
@@ -12,7 +12,7 @@ import {
 } from './evaluation-helpers.js';
 import type { Violation } from '../types/evaluation.js';
 import type { TestResult, CacheStatistics } from '../types/test-result.js';
-import type { ProgrammaticViolation } from '../../src/programmatic/types';
+import type { ProgrammaticViolation } from '@/validation/types';
 
 /**
  * Generates a markdown report from evaluation results
