+++ packages/@n8n/ai-workflow-builder.ee/src/validation/evaluators/connections.ts
