+++ packages/@n8n/ai-workflow-builder.ee/src/validation/utils/is-tool.ts
