+++ packages/@n8n/ai-workflow-builder.ee/src/validation/utils/score.ts
