+++ packages/@n8n/ai-workflow-builder.ee/src/workflow-state.ts
@@ -2,7 +2,7 @@ import type { BaseMessage } from '@langchain/core/messages';
 import { HumanMessage } from '@langchain/core/messages';
 import { Annotation, messagesStateReducer } from '@langchain/langgraph';
 
-import type { ProgrammaticEvaluationResult } from './programmatic/types';
+import type { ProgrammaticEvaluationResult } from '@/validation/types';
 
 import type { SimpleWorkflow, WorkflowOperation } from './types/workflow';
 import type { ChatPayload } from './workflow-builder-agent';
